The big cutover to ORM begins! What's happening here: 1. the metadata.py file is being updated to use the ORM 2. conflicting models are being sunset and/or quarantined for this PR 3. CRUD accessors stay in metadatastore but are now managed behind the scenes by the ORM
This is going to break a lot of things (which is goodTo get unbroken:
1. update the tests to no longer be aware of the backend configs
2. update the code to same
3. remove all the SQLModel and deprecated backend code
4. document (loom) how the ORM works, how to create migrations, how
to traverse the ORM tree etc etc.

Strategy here should be to merge this into a long-running branch and start CI against it,
then keep pulling main into it until we're ready for a major release (this will be a major).

Configs will be extremely thin after this PR. We should be
set up to move docker dev to a single stack and docker quickstart to a single image.